Refactored basic code into health-mon base class.
